feat: refactor chat module by adding new classes for user caching, message rendering, WebSocket handling, and UI management
