+++ frontend/src/ts/chat/index.ts
@@ -1,0 +1,7 @@
+// Export all chat-related modules
+export { Chat } from "./Chat.js";
+export { UserCache, type CachedUser } from "./UserCache.js";
+export { MessageRenderer } from "./MessageRenderer.js";
+export { WebSocketHandler } from "./WebSocketHandler.js";
+export { ChatActions } from "./ChatActions.js";
+export { ChatUI } from "./ChatUI.js";
